fix auth code callback
