+++ frontend/src/hooks/useCallbackURL.ts
@@ -1,5 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
-import { getCallbackURL } from "@/api/auth";
+import { getCallbackURL, sendCode } from "@/api/auth";
+import { useEffect, useRef } from "react";
+import { useState } from "react";
 
 /**
  * Provides a callback URL for authentication.
@@ -18,4 +20,39 @@ export function useCallbackURL() {
     },
     retry: false,
   });
+}
+
+/**
+ * Sends an authentication code after returning from the auth callback URL.
+ * 
+ * @param code - The code to send.
+ * @param onSuccess - The function to call on success.
+ * @param onError - The function to call on error.
+ * @returns The loading state.
+ */
+export function useAuthCallback(code: string | null, onSuccess: () => void, onError: (err: Error) => void): boolean {
+  const [isLoading, setIsLoading] = useState(true);
+  const hasRun = useRef(false);
+
+  useEffect(() => {
+    if (hasRun.current) return; // prevent multiple executions
+    hasRun.current = true;
+
+    // Error if no code provided
+    if (!code) {
+      onError(new Error("No login code provided"));
+      return;
+    }
+
+    sendCode(code)
+      .then(() => {
+        onSuccess();
+        setIsLoading(false);
+      })
+      .catch((err: Error) => {
+        onError(err);
+      });
+  }, [code]);
+
+  return isLoading;
 }
